+++ chevah/compat/nt_filesystem.py
@@ -171,7 +171,7 @@ class NTFilesystem(PosixFilesystemBase):
         """
         letter, _ = os.path.splitdrive(path)
         if letter.strip(':').lower() not in self._allowed_drive_letters:
-            message = 'Bad drive letter "%s" for %s' % (letter, path)
+            message = u'Bad drive letter "%s" for %s' % (letter, path)
             raise OSError(errno.EINVAL, message.encode('utf-8'))
 
     def getSegmentsFromRealPath(self, path):
@@ -187,7 +187,8 @@ class NTFilesystem(PosixFilesystemBase):
         path = os.path.abspath(path)
 
         if self._avatar.lock_in_home_folder:
-            # Locked filesystem have no drive.
+            self._checkChildPath(self._getRootPath(), path)
+            # Locked filesystems have no drive.
             tail = path[len(self._getRootPath()):]
             drive = ''
         else:
@@ -412,6 +413,20 @@ class NTFilesystem(PosixFilesystemBase):
         except OSError:
             return False
 
+    def deleteFile(self, segments, ignore_errors=False):
+        """
+        See `ILocalFilesystem`.
+        """
+        try:
+            return super(NTFilesystem, self).deleteFile(
+                segments, ignore_errors=ignore_errors)
+        except OSError, error:
+            # This should also catch WindowsError and re-raise as OSError.
+            error_number = error.errno
+            if error_number == errno.EINVAL:
+                error_number = errno.ENOENT
+            raise OSError(error_number, error.strerror)
+
     def deleteFolder(self, segments, recursive=True):
         """
         See `ILocalFilesystem`.
@@ -625,7 +640,10 @@ class NTFilesystem(PosixFilesystemBase):
         See `ILocalFilesystem`.
         """
         if self.isLink(segments):
-            target_segments = self.readLink(segments)
-            return self.exists(target_segments)
+            try:
+                target_segments = self.readLink(segments)
+                return self.exists(target_segments)
+            except CompatError:
+                return False
         else:
             return super(NTFilesystem, self).exists(segments)