+++ release-notes.rst
@@ -1,5 +1,26 @@
-Release notes for chevah.empirical
-==================================
+Release notes for chevah.compat
+===============================
+
+
+0.18.1 - 24/03/2014
+-------------------
+
+* LocalFilesystem.exists() now returns false on Windows for broken links.
+
+
+0.18.0 - 24/03/2014
+-------------------
+
+* Raise CompatError in getSegmentsFromRealPath if path is outside of home
+  folder.
+
+
+0.17.1 - 20/03/2014
+-------------------
+
+* Update build system to latest buildbot.
+* Convert WindowsError from deleteFile into OSError and convert error code
+  for file not found.
 
 
 0.17.0 - 04/03/2014